删除多余参数，修改成将原始的数据送入系统

+++ include/vio_manager.h
@@ -74,13 +74,13 @@ public:
     ~VioManager();
 
     void InitVioManager(void (*PublishVIO)(
-                        uint64_t timestamp,
+                        double timestamp,
                         const Eigen::Vector3d &p,
                         const Eigen::Quaterniond &q,
                         float covVx, float covVy, float covVz,
                         uint8_t resetFlag,float rate1, float rate2),
                         void (*PublishPoints)(
-                        uint64_t timestamp, 
+                        double timestamp, 
                         const std::vector<int> &curr_init_ids,
                         const std::vector<Eigen::Vector2d> &curr_init_obs,
                         const std::vector<Eigen::Vector3d> &curr_init_pts));
